Added linked list implementation, set and resize
Basically I read between the lines of the assignment which seemed to
want me to handle collisions with linked lists even though personally I
would prefer nested arrays in this case but whatever I am nothing if not
a good student so here we go. I imported the linkedList stuff from the
previous project and made small tweaks for it to work with key value
pairs. I also added the set method per the assignment, which allows us
to add key value pairs to the appropriate bucket, overwrites if a
duplicate key is being used, or adds it to the head of the linked list
if the keys are different with the same hash value. After something is
added to the array, we check to see if it needs to be resized. If it
does, then resize is dynamically run, which expands the capacity and
goes through the rehashing process by using the old data from the
array we had before expanding the capacity. What an ass pain!