add remove method, update list methods
The remove method would have been way too verbose without relying on the
methods from linkedList, so I updated those methods to make it work more
smoothly. I also added a getIndex method to the linkedList that returns
the index of the given key so that said key can be removed.